Docking API Added
Many, many errors in the Application Framework, but I have added in the Docking API to this project, as I believe they are related.

Since this project is concerned with separating the application plumbing from the display and business logic of an application, I figured this is where the Docking API needs to reside also. Therefore, I have added the Swing Docking Framework API into the Swing Application Framework Project.

Now, we can work on both of these frameworks simultaneously and in conjunction with each other. This will greatly reduce the required code for Northwind Traders Basic Edition, as well as all future projects PekinSOFT Systems may have.

-SC