Work Continues on Swing Application Framework
I am making slow-but-sure progress on the Swing Application Framework. Many classes still have errors, though some of the errors that were present have been eliminated by implementing other referenced classes. The errors remaining are referenced classes that have yet to be written.

This commit and push is to prepare to leave in the morning. I will probably not do another push for a little while. I just need to get what I have accomplished onto the repository so that I can clone it to the laptop I take in the truck with me.

@knathan is currently working on improving the Swing Docking Framework, which I have refactored into this project in an effort to keep this all together, since both libraries are related to the application plumbing.

That is about it for now. More later...

-SC