Weekly Update
This is just a weekly update...there are still errors as there are still
classes that need to be created. Work is progressing decently, but I am
pausing the work on this for the weekend because I am home and need to
get the website up and working properly for @jkovalsky to be able to 
test the NLS in a "live" environment.

-SC