Added View and FrameView Classes
I have added the `View` and `FrameView` classes for @knathan for working
on the Swing Docking Framework. These classes will, allegedly, make the
Docking Framework easier and better.

Per the JavaDoc on the `View` class:

> The advantage of Views over just configuring a JFrame directly, is that
> a View is more easily moved to an alternative top level container, like
> a docking framework.

Due to this information coming to my attention, I figured that I needed
to get these classes done now, so that they could be used within the
Swing Docking Framework library, instead of the `JInternalFrame`s that
we are currently working with.

The `View` class is extending the `AbstractBean` class, and the 
`FrameView` class is extending the `View` class. Both of these classes
are 100% complete and ready to be used in the Swing Docking Framework
library at this time. This is the whole reason for this commit and push,
so that these two classes are available to @knathan.

There are still a lot of classes that need to be written, as well as a
bunch that still need to be completed, within the Swing Application
Framework library, so the library as a whole is not ready at this time,
but work is progressing nicely.

-SC