Swing Application Framework Complete!
No more errors, so it must be done...

All files are included in this push, as I have completed all of the classes and
refactored the entire library back to `org.jdesktop.application`, with various subpackages.

The library should now be usable, but I have yet to test it out. It would be really nice if we could make this into a NetBeans Plugin, the way the original Swing Application Framework was, then we could use it all through the NetBeans interface the way it was then. Well, I guess that could be my next task...

Anyway, @knathan, you need to pull the library one more time so that you have the final version to work on the Swing Docking Library.

-SC