JavaDoc Housekeeping
I just performed some JavaDoc housekeeping, as I had errors and warnings with the JavaDoc compiler. I have fixed all errors, but not the warnings, as most of these are for missing JavaDoc comments, missing `@param` descriptions, missing `@return` descriptions, etc. I can work on those in the future. Right now, I just wanted to get the errors corrected.

Also, I refashioned the JavaDoc code from using `{@code *}` tags, for in-line code, and `<code></code>` tags, for code blocks. I switched over to using the markdown tags so that way the in-line code words will not be on their own line and the code blocks will be syntax-highlighted for Java. This is possible because I am not using the JavaDoc compiler for creating the API documentation, but am using `doxygen` instead.

-SC