Application Framework Review
I have reviewed the Application Framework for my errors. I believe that I have found and fixed them all. I am going to start testing it now.

part of the issue was that I had incorporated the PekinSOFT Utilities into the Framework, but had not actually done the work to integrate them. Currently, as of this push, the PekinSOFT Utilities are there, but not connected to the Framework. I will get those going once I make sure the Framework is working properly.

Also, I started creating the `DockingFrameApplication` class that I had discussed in email. I stopped working on it because I was more interested in fixing the Framework than adding to it at this point. However, I just wanted to note that it is started.

-SC